Разделение файла модели-представления на два
