+++ MailSender/ViewModels/MainWindowViewModel.Commands.cs
@@ -1,0 +1,109 @@
+﻿using System.Linq;
+using System.Windows;
+using System.Windows.Input;
+using MailSender.Infrastructure.Commands;
+using MailSender.Models;
+
+namespace MailSender.ViewModels
+{
+    partial class MainWindowViewModel
+    {
+        #region Команды
+
+        #region CreateNewServerCommand
+
+        private ICommand _CreateNewServerCommand;
+
+        public ICommand CreateNewServerCommand => _CreateNewServerCommand
+            ??= new LambdaCommand(OnCreateNewServerCommandExecuted, CanCreateNewServerCommandExecute);
+
+        private bool CanCreateNewServerCommandExecute(object p) => true;
+
+        private void OnCreateNewServerCommandExecuted(object p)
+        {
+            // Основное действие, выполняемое командой, описывается здесь!!!
+
+            MessageBox.Show("Создание нового сервера!", "Управление серверами");
+        }
+
+        #endregion
+
+        #region EditServerCommand
+
+        private ICommand _EditServerCommand;
+
+        public ICommand EditServerCommand => _EditServerCommand
+            ??= new LambdaCommand(OnEditServerCommandExecuted, CanEditServerCommandExecute);
+
+        private bool CanEditServerCommandExecute(object p) => p is Server || SelectedServer != null;
+
+        private void OnEditServerCommandExecuted(object p)
+        {
+            var server = p as Server ?? SelectedServer;
+            if (server is null) return;
+
+            MessageBox.Show($"Редактирование сервера {server.Address}!", "Управление серверами");
+        }
+
+        #endregion
+
+        #region DeleteServerCommand
+
+        private ICommand _DeleteServerCommand;
+
+        public ICommand DeleteServerCommand => _DeleteServerCommand
+            ??= new LambdaCommand(OnDeleteServerCommandExecuted, CanDeleteServerCommandExecute);
+
+        private bool CanDeleteServerCommandExecute(object p) => p is Server || SelectedServer != null;
+
+        private void OnDeleteServerCommandExecuted(object p)
+        {
+            var server = p as Server ?? SelectedServer;
+            if (server is null) return;
+
+            Servers.Remove(server);
+            SelectedServer = Servers.FirstOrDefault();
+
+            //MessageBox.Show($"Удаление сервера {server.Address}!", "Управление серверами");
+        }
+
+        #endregion
+
+        #region Command SendMailCommand - Отправка почты
+
+        /// <summary>Отправка почты</summary>
+        private ICommand _SendMailCommand;
+
+        /// <summary>Отправка почты</summary>
+        public ICommand SendMailCommand => _SendMailCommand
+            ??= new LambdaCommand(OnSendMailCommandExecuted, CanSendMailCommandExecute);
+
+        /// <summary>Проверка возможности выполнения - Отправка почты</summary>
+        private bool CanSendMailCommandExecute(object p)
+        {
+            if (SelectedServer is null) return false;
+            if (SelectedSender is null) return false;
+            if (SelectedRecipient is null) return false;
+            if (SelectedMessage is null) return false;
+            return true;
+        }
+
+        /// <summary>Логика выполнения - Отправка почты</summary>
+        private void OnSendMailCommandExecuted(object p)
+        {
+            var server = SelectedServer;
+            var sender = SelectedSender;
+            var recipient = SelectedRecipient;
+            var message = SelectedMessage;
+
+            var mail_sender = _MailService.GetSender(server.Address, server.Port, server.UseSSL, server.Login, server.Password);
+            mail_sender.Send(sender.Address, recipient.Address, message.Subject, message.Body);
+
+            Statistic.MessageSended();
+        }
+
+        #endregion
+
+        #endregion
+    }
+}
